First commit of EAU after changing the names
This is the first commit of EAU. It changes the names of everything
from the ELN to EAU name. For the most part.

All of the textures and models are working from the normal game,
except for the arc furnace (new texture and code in the works),
and the sterling engine model is not yet imported.

The rest of the removals are intentional, and I have left models
in place in mostly a just-in-case situation.

The game compiles and runs, there are no missing textures. A lot
of work to do on the textures, the background "voltage tier colors",
as well as some work to be done with the language files.

Some stuff is converted to Kotlin.

PR away! :)